Delete conflicting JS.jpg duplicate file
The file JS.jpg and js.jpg cannot co-exist in the same directory on a case-insensitive filesystem such as NTFS on Windows. This results in a warning when cloning the repo onto such file system. The following warning message is displayed.

warning: the following paths have collided (e.g. case-sensitive paths
on a case-insensitive filesystem) and only one from the same
colliding group is in the working tree:
  'images/JS.jpg'
  'images/js.jpg'